Refactor eval.rs into submodules; linear scan matching
- Split 2500-line eval.rs into 8 focused submodules under eval/:
  accumulate, evaluate, iou, summarize, tide, confusion, types, mod
- Extract precision_recall_curve as a free function shared between
  accumulate and tide_errors (eliminates per-call Vec alloc in tide)
- Hoist MetricDef struct and metrics_bbox_segm/metrics_kp builders to
  module level (were local to summarize() function body)
- Replace sort_by_iou + Vec<Vec<usize>> pre-sort with a direct linear
  scan: eliminates 2×D Vec allocations per (image, category) pair;
  correct and faster for typical COCO (≤5 GTs/cat)
- Single-pass unzip for GT collection in compute_segm_iou_static and
  compute_bbox_iou_static (eliminates double iteration over gt_ids)
- Tighten visibility: EvalImgContext → pub(super); accumulate mod →
  pub(super) so tide.rs can access precision_recall_curve as a sibling
- Re-export EvalShape from crate root
- Fix broken rustdoc intra-doc links in ConfusionMatrix and TideErrors

Parity verified: bbox max diff 7.05e-05, segm 1.01e-04, kpts 1.67e-15
